refactor(build-logic): move to build-logic compiled plugins, separate integration-test module for TLS/NatsClient integration
