+++ build-logic/convention/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+	`kotlin-dsl`
+}
+
+dependencies {
+	implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:${libs.versions.kotlin.stdlib.get()}")
+}
