+++ build-logic/convention/src/main/kotlin/EnsureNatsHarnessTask.kt
