+++ build-logic/convention/src/main/kotlin/NatsServerDaemonService.kt
