+++ build-logic/convention/src/main/kotlin/natskt.kmp.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+	kotlin("multiplatform")
+}
