+++ build-logic/settings.gradle.kts
@@ -1,0 +1,16 @@
+rootProject.name = "build-logic"
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+
+dependencyResolutionManagement {
+	repositories {
+		mavenCentral()
+		gradlePluginPortal()
+	}
+	versionCatalogs {
+		create("libs") {
+			from(files("../gradle/libs.versions.toml"))
+		}
+	}
+}
+
+include(":convention")
