--- buildSrc/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-	`kotlin-dsl`
-}
-
-repositories {
-	mavenCentral()
-	gradlePluginPortal()
-}
