+++ core/common/build.gradle.kts
@@ -1,32 +1,10 @@
-@file:OptIn(ExperimentalWasmDsl::class)
-
-import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
-
 plugins {
-    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.natskt.kmp)
 }
 
 kotlin {
     explicitApi()
-
-    jvm()
-    js {
-        browser()
-        nodejs()
-    }
-
-    wasmJs {
-        browser()
-        nodejs()
-    }
-
-    iosArm64()
-    iosSimulatorArm64()
-
-    linuxX64()
-    linuxArm64()
-
-    macosArm64()
+    allTargets()
 
     sourceSets {
         val commonMain by getting {
