+++ gradle/libs.versions.toml
@@ -74,6 +74,7 @@ ktor-server-engine-netty = { module = "io.ktor:ktor-server-netty", version.ref =
 ktor-server-content-negotiation = { module = "io.ktor:ktor-server-content-negotiation", version.ref = "ktor"}
 
 [plugins]
+natskt-kmp = { id = "natskt.kmp" }
 kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin-stdlib" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin-stdlib" }
