+++ jetstream/build.gradle.kts
@@ -4,14 +4,14 @@ import java.time.Duration
 import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 
 plugins {
-	alias(libs.plugins.kotlin.multiplatform)
+	alias(libs.plugins.natskt.kmp)
 	alias(libs.plugins.kotlin.serialization)
 }
 
 kotlin {
     explicitApi()
+    allTargets()
 
-    jvm()
     js {
         browser {
 			testTask {
@@ -48,14 +48,6 @@ kotlin {
 			}
 		}
     }
-
-    iosArm64()
-    iosSimulatorArm64()
-
-    linuxX64()
-    linuxArm64()
-
-    macosArm64()
 
     sourceSets {
         commonMain.dependencies {
