+++ settings.gradle.kts
@@ -2,6 +2,7 @@ rootProject.name = "natskt"
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 pluginManagement {
+    includeBuild("build-logic")
     repositories {
         mavenCentral()
         gradlePluginPortal()
